requirements/development.txt
requests==2.21.0
beautifulsoup4==4.6.3
celery==4.1.1
redis==2.10.5
walrus==0.4.1
python-slugify
